Crate wasmi_validation

Modules

context

func

stack

util

Structs

Error

PlainFuncValidator

A function validator that just validates modules and produces no result.

PlainValidator

A module validator that just validates modules and produces no result.

Constants

DEFAULT_MEMORY_INDEX

Index of default linear memory.

DEFAULT_TABLE_INDEX

Index of default table.

LINEAR_MEMORY_MAX_PAGES

Maximal number of pages that a wasm instance supports.

Traits

FuncValidator

Validator

Functions

validate_memory

validate_module